pyproject.toml
[project]
name = "tone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.0.0",
    "cryptography>=41.0.0",
    "fastapi>=0.104.0",
    "loguru>=0.7.0",
    "psycopg2-binary>=2.9.0",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.24.0",
]
